requirements.txt
pyyaml
feedparser
requests
beautifulsoup4
notion-client
